# Pipeline a bunch of sync operations