Adds PG pipelining to a bunch of sync operations in the hope it could speed things up.

I tested this out on staging by pulling the latest base release on each version.

main: 
* 28.4 s (cpu), 25.5 s (system)

this branch:
* 29.1 s (cpu), 25.3 s (system)

No perceptible difference AT ALL, so that suggests query latency isn't any significant bottleneck for pull, bummer.